+++ hutool-core/src/main/java/cn/hutool/core/lang/PatternPool.java
@@ -61,6 +61,8 @@ public class PatternPool {
 	public final static Pattern PLATE_NUMBER = Pattern.compile(
 			//https://gitee.com/loolly/hutool/issues/I1B77H?from=project-issue
 			"^(([京津沪渝冀豫云辽黑湘皖鲁新苏浙赣鄂桂甘晋蒙陕吉闽贵粤青藏川宁琼使领][A-Z](([0-9]{5}[ABCDEFGHJK])|([ABCDEFGHJK]([A-HJ-NP-Z0-9])[0-9]{4})))|" +
+			//https://gitee.com/loolly/hutool/issues/I1BJHE?from=project-issue
+			"([京津沪渝冀豫云辽黑湘皖鲁新苏浙赣鄂桂甘晋蒙陕吉闽贵粤青藏川宁琼使领]\\d{3}\\d{1,3}[领])|" +
 			"([京津沪渝冀豫云辽黑湘皖鲁新苏浙赣鄂桂甘晋蒙陕吉闽贵粤青藏川宁琼使领][A-Z][A-HJ-NP-Z0-9]{4}[A-HJ-NP-Z0-9挂学警港澳使领]))$");
 
 	// -------------------------------------------------------------------------------------------------------------------------------------------------------------------