+++ hutool-core/src/test/java/cn/hutool/core/lang/ValidatorTest.java
@@ -134,5 +134,6 @@ public class ValidatorTest {
 	@Test
 	public void isPlateNumberTest(){
 		Assert.assertTrue(Validator.isPlateNumber("粤BA03205"));
+		Assert.assertTrue(Validator.isPlateNumber("闽20401领"));
 	}
 }